requirements.txt
mandrill==1.0.60
mysql-connector-python-rf==2.1.3
python-dotenv==0.6.0